+++ transaction/transaction.go
@@ -1,8 +1,8 @@
 package transaction
 
 type Transaction struct {
-	Hash string
-	Gas int
-	FeePerGas float32
+	Hash      string
+	Gas       int
+	FeePerGas float64
 	Signature string
 }
